requirements-vercel.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
langchain==0.0.335
sentence-transformers==2.2.2
chromadb==0.4.18
google-generativeai==0.3.1
python-dotenv==1.0.0